Abs expression moved to Operator Nodes and minor fixes
